+++ prefect_aws/ecs.py
@@ -105,6 +105,7 @@ Examples:
 """
 import copy
 import difflib
+import json
 import logging
 import pprint
 import sys
@@ -257,7 +258,7 @@ class ECSTask(Infrastructure):
         task_role_arn: An optional role to attach to the task run.
             This controls the permissions of the task while it is running.
         task_customizations: A list of JSON 6902 patches to apply to the task
-            run request.
+            run request. Can be a string to support deployment cli overrides.
         task_start_timeout_seconds: The amount of time to watch for the
             start of the ECS task before marking it as failed. The task must
             enter a RUNNING state to be considered started.
@@ -434,7 +435,7 @@ class ECSTask(Infrastructure):
             "task while it is running."
         ),
     )
-    task_customizations: JsonPatch = Field(
+    task_customizations: Union[List[Dict], JsonPatch, str] = Field(
         default_factory=lambda: JsonPatch([]),
         description="A list of JSON 6902 patches to apply to the task run request.",
     )
@@ -538,14 +539,16 @@ class ECSTask(Infrastructure):
 
     @validator("task_customizations", pre=True)
     def cast_customizations_to_a_json_patch(
-        cls, value: Union[List[Dict], JsonPatch]
+        cls, value: Union[List[Dict], JsonPatch, str]
     ) -> JsonPatch:
         """
         Casts lists to JsonPatch instances.
         """
+        if isinstance(value, str):
+            value = json.loads(value)
         if isinstance(value, list):
             return JsonPatch(value)
-        return value
+        return value  # type: ignore
 
     class Config:
         """Configuration of pydantic."""